Removed the hardcoded virtual address based purposeful mis adventure.
Added offsets for the Access, Status and Data Fifos for all the 5 localities
of the TPM.

Added bit definitions for some of the bits in the Access register/port/iomem.

Added a tpm_init function which tries to relinquish control of all the
localities, just in case. Then it requests access to locality 0, so that
tpm_dump_info can get the vendor id.