Updated the module version number as well as
FIXED a SIMPLE yet POWERFUL goofup in my padconf setup logic.
The earlier stupid? logic would have allowed the module to run
only once. This is not desired normally, however it is a nice
side effect to have at a later stage, when we want to control
has to how many times the module should be allowed to be loaded,
in a simple and stupid way.