Noticied that if I increase the property count in GetCap command structure, then the TPM seems to send the subsequent property tags starting from the one specified in the GetCap commands Property field.
So added a new tpm library call called tpm_getcap_ptfixed, which
currently starts at the 1st Fixed Property tag i.e

TPM_PT_FAMILY_INDICATOR, for which I got the response string as 2.0
Well actually

01 		/* MoreData Yes */
00 00 00 06	/* TPM Properties */
00 00 00 01	/* I assume 1 property follows */
00 00 01 00	/* Property tag: PT_FAMILY_INDICATOR */
32 2e 30 00	/* 2.0 */

Currently I have set the property count to only 1. so I got a response
buffer with size 0x1b, with the above result.

Next I will increase the property count in GetCap command to 2 and will
check the size it should be 0x1b+1*4*2 = 0x23 (i.e 4 bytes for next PT marker
and 4 bytes for the value corresponding to that PT)

In which case after this I will increase the property count to 45, so that
I get all the PT_FIXED properties in one shot. It should require a buffer
of 0x1b+44*4*2 = 0x17b.

********* IMPORTANT *************
NOTE: I had forgotten to show the 01 /* MoreData Yes */ Byte field in the
response following the response header in the previous git commit message.
This field was in the actual response which I got, but I forgot to mention
the same in the git commit message. That field was also one of things
along with property count (in command structure) to help me experiment wrt
getting more fields as mentioned above.