Added the tpm_pcr_extend logic.
As part of the same, also understood the basic and the simplest
The password based authorisation and session(rather which is implicit
in case of password auth based sessions and inturn the commands where
it is used).

This should help with the next step of locking the tpm using the
password based auth for now. Later will have to add more complex
authorisation mechanisms.

NOTE4PREVCOMMIT: When I added the tpm_pcr_read_all, I also verified
that the 24 PCRs match the reset value specified by the
PC Client Specific PTP for TPM2 i.e

00-16 = All 0s
17-22 = All 1s
23    = All 0s

Now after adding the tpm_pcr_extend logic with a input digest of
0123456789012345678901234567890123456789
Got a proper pcr extended value of
6f5b5b585ee60ee1053c12e42ac5f2876a34ab22