Added the tpm_hierarchy_changeauth logic which currently
a) changes the Owner password, provided it is being done
   for the first time on the TPM.

As expected, the 1st time it was run, the password seems to have
been successfully changed, because the 2nd time it was run, I got
a session error with authorisation failure without DA implications

ResponseCode = 0x9a2
Error code within it is 0x22

TODO: Currently I am blindly calling the logic to print the
authorisation structure in the response buffer. But if there is
a error, then there wont be a authorisation structure in respbuf.
I have to fix this