Renamed the Original HierarchyChangeAuth Hand Built CommandBuffer to make the name more approriate
Added a new HierarchyChangeAuth command buffer, which can be used
anytime, because it has space for specifies a Non Null password
as the current password, as well as specify a new password.

Update the tpm_hierarchy_changeauth command to accept the command
buffer, so that it can be used with the initial or anytime command
buffer.

However I noticied that input buffer size is showing has 8 bytes,
and so inturn the commands are failing, I have to cross check,
what mistake I might have done, here.

Also added some missing function declaration to tpmk_lib.h file.