On Reading the TPM2.0 Library Spec Part-1 architecture, it seems to indicate that
a) PlatformAuth is always reset to EmptyAuth/NullAuth when ever
the TPM is Reset or Restarted (definitely on a power cycle)

b) OwnerAuth and EndorsementAuth are only reset on a TPM Clear

c) LockOutAuth only if prev LockOutAuth is given or even on TPM Clear
Have to crosscheck this once again.

So I have changed the logic so that

Each time this software is run, by default only the platformAuth is
newly set. But if the software is being run on a platform/system
for the very first time (the software requires to be recompiled
in this case, with the appropriate constant defined) then all 4
control domains passwords are set to a new value.

At same time the existing password set by this system, is rechecked
for each of the control domains always.

In order to achieve this, hierarchy_changeauth has been suitably
updated.

UPDATE + TODO:

*** I ***

1) Have verified that indeed the PlatformAuth is being reset
to a NulLAuth each time the system is power cycled.

2) I am able to changeauth for Platform, Owner and Endorsement
domains, using the previous authPassword I had set.

3) For some reason even thou I think I am sending the proper
lockoutAuthPassword, which I had set previoulsy, it doesnt seem
to be allowing changing the auth password for lockoutAuth. Have
to check the spec once again, to see

3.1) if changing lockoutAuth requires any additional authorisation
other than the existing lockoutAuth

3.2) As previously, as well as even now, for testing purpose I had
purposefully triggered wrong password based changeauth, if the
lockout mechanism of the tpm is what is stopping me from changing
the lockoutAuth, even with a valid lockoutAuth (The spec does talk
about this corner case, as a possibility for a TPM).

3.3) Have to look at the lockout related parameters in the tpm, and
inturn clear the lockout state in the tpm, which I had previously
triggered. As its automatic clearing may take 72 hours of runtime
or so (If I remember the spec clearly).

*** II ***

Have to look at how the tpm clock is progressing singularly forward
when ever it has power available to it. And what is the granularity
of time required for the tpm to update its internal clock / time
passing into its NV storage associated with clock.