Registering as a char device with Major Number 200. As I have used the major number pretty arbitrarily, it may clash in a given system, in which case the major number will have to be changed.
Also added some prints to help identify when any user process
opens or closes the device.

Added a python script, which accepts a hex string as a argument
and passes the same to the tpm as a command and inturn reads the
response and prints both the command and response. It also takes
care of creating the char device file in the system, obviously
will succeed only if run as root.

In turn using this python script python-tpmif.py I have created few
shell scripts to

a) get the manufacturer using getcap command

b) call readclock