Cleaned up the Module parameters to allow better fine grained control over what occurs during module loading.
a) TpmClear can be enabled or disabled

b) Auths can be initialised (i.e assumes Null / Empty Auth as current)
   or Not.

   0 to not initialize any auths
   1 to initialize only Platform Auth, which gets cleared on each
     power cycle (this is what is normally required)
   4 or more to initialise all the Auths (i.e Platform, Owner,
     Endorsement and Lockout). (This is only required for a
     new TPM, or after a TPM Clear).

c) Auths can be verified (i.e assume a existing non null auth)

NOTE1: As of now all auths are initialised to password auths.
Also assumes that the existing auth is a password auth, which is
set by this software itself.

NOTE2: A TPMClear may clear only Owner, Endorsement and LockOut
Auth potentially, i.e Platform Auth may not be changed by it.
Have to verify this once in the specs.

UPDATE: Previously I had missed out a section of response codes
at the end of the response code list, which corresponded to the
situation where severity is set as warning. So I was treating
the TPM_RC_LOCKOUT error response which I was getting from TPM
for LockoutAuth related commands (including potentially for
TpmClear with LockoutAuth, have forgotten) as TPM_RC_EXCLUSIVE
which was not making sense.

Once I seem to have cleared even LockOut Properly, but I have
to verify / experiment more on this.